refactor code: separate ColorUtil out from RPTUtil #1182
