+++ accessibility-checker-engine/test/v2/checker/accessibility/rules/IBMA_Color_Contrast_WCAG2AA_ruleunit/color_with_transparency.html
@@ -14,6 +14,13 @@
         </div>
       </div>
       <div style="background: rgb(33, 33, 33);">
+        <div style="background-color: hsla(0, 0%, 100%);">
+          <div style="color: white;">
+            This should not fail, but does
+          </div>
+        </div>
+      </div>
+      <div style="background: rgb(33, 33, 33);">
         <div style="background-color: hsla(0, 0%, 100%, 0.2);">
           <div style="color: white;">
             This should not fail, but does
